requirements.txt
# Web Framework
flask==3.1.1
jinja2==3.1.6
click==8.2.2
itsdangerous==2.2.0
markupsafe==3.0.2
blinker==1.9.0
werkzeug==3.1.3

# Environment & Configuration
python-dotenv==1.1.1

# HTTP Requests
requests==2.32.4

# AI APIs
assemblyai==0.42.1
google-genai==1.29.0

# Optional: Alternative frameworks (if needed)
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
